requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
slack_bolt==1.27.0
slack_sdk==3.39.0
urllib3==2.6.3
